requirements.txt
atlassian-python-api>=3.41.4
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0
html2text>=2020.1.16 